M1/external-model-google-genai-py/requirements.txt
google-genai
python-dotenv
tiktoken
